requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
openpyxl
plotly>=5.18.0
statsmodels
faker
pyarrow>=14.0.0
wordcloud
matplotlib>=3.5.0
duckdb>=0.9.0
datasets>=2.14.0
huggingface-hub>=0.17.0
supabase
python-dotenv
tqdm
